Remove pointless reallocation when refreshing cache.
previously, in order to save performance, rather than setting
every value in the cache bytearray to 0, a new one was allocated
and assigned in place of the old one.
This caused issues with from imports, and didn't end up saving
performance, since the default color is white, so iterating over
every element of the cache was needed anyway.