Redis: switch to a Redis powered cache
This commit change our cache method from a (broken as it was lacking IPC) bytearray to Redis.

It adds a new meta table cache_state in the database storing three new fields:
- last_modified: when the pixel_history table was last modified. It is kept up-to-date by a trigger, meaning that even manual access to the database will update it.
- last_updated: when the Redis cache has been last updated. Manually bumped.
- sync_lock: whenever an instance is already syncing the cache. It works in a spinlock-like manner.

Before any transaction, the new Canvas class will check if last_modified > last_updated. If so, a synchronisation will start.
Before actually attemping the synchronisation, the client will spin the sync_lock. If it didn't acquire it, it will consider that another client is already syncing it and will wait until the lock is released before moving on.

Do note that this may deadlock the service where it is indefininely waiting for the lock to be released but due to an error it never will (such as the app being killed while syncing).
I tried to add a timeout where after some point waiting it will consider the lock as stalled and release it. My problem was that every worker tried to release it at the same time. I tried to mitigate that by shifting the waiting time by a specific value (I choose the PID) adn restarting the whole application after releasing the lock.
Although due to the inner (undocumented) working of the Docker image, I couldn't get it to work. Right now if something like that happens, somebody with database access will have to clear the row.

As for the syncing process itself, the current cache is considered as invalid and is entirely replaced by data taken from the current_pixel view.
In order to optimize Redis writes, one key-value is used per line in the format "canvas-line-$line". At the end of the sync, last_updated takes the value of now()

Two operations are provideds added to pixel_history, it is inserted to the cache line and last_updated is modified to store now().
2. Get pixels: Every cache line is aggregated into one big bytearray and returned to the caller.

/set_pixel and /post_pixels have also be updated to use those new method. A global canvas instance is stored in pixels.pixels as it is impossible to pass the Redis connection around otherwise.