Skip canvas valid check on ban
When banning someone previously, if another edit came in after the pixel_history
is marked as deleted, the cache state would be set as up to date, so when the
sync_cache() call was ran, it wouldn't refresh the cache.

This commit adds a kwarg to froce a cache refresh when needed.